[#7] Refactor: extract URL construction logic as makeCategoryURL()
