+++ SideDishApp/SideDishApp/Networking/NetworkManager.swift
@@ -8,47 +8,35 @@
 import Foundation
 import Alamofire
 
-enum NetworkError: Error {
-    case noData
-    case wrongBaseEndPoint
-    case wrongEndPoint
-}
-
-extension NetworkError: LocalizedError {
-    var errorDescription: String? {
-        switch self {
-        case .noData: return NSLocalizedString("\(self)", comment: "No Data")
-        case .wrongBaseEndPoint: return NSLocalizedString("\(self)", comment: "Wrong Base Endpoint")
-        case .wrongEndPoint: return NSLocalizedString("\(self)", comment: "Wrong EndPoint")
-        }
-    }
-}
-
 struct NetworkManager {
-    let baseEndPoint = "https://api.codesquad.kr/onban"
     
     func fetchProducts(of type: ProductType, completion: @escaping (Result<[Product], NetworkError>) -> Void) {
         
-        guard var component = URLComponents(string: baseEndPoint) else {
-            return completion(.failure(.wrongBaseEndPoint))
-        }
-        
-        component.path += "/\(type)"
-        
-        guard let url = component.url else {
-            return completion(.failure(.wrongEndPoint))
-        }
-        
-        AF.request(url)
-            .validate()
-            .responseDecodable(of: Response.self) { AFResponse in
-                
+        do {
+            let url = try makeCategoryURL(of: type)
+            
+            AF.request(url).validate().responseDecodable(of: Response.self) { AFResponse in
                 guard let products = AFResponse.value?.body else {
                     return completion(.failure(.noData))
                 }
-                
                 completion(.success(products))
             }
+        } catch {
+            SystemLog.fault(error.localizedDescription)
+        }
+    }
+    
+    private func makeCategoryURL(of type: ProductType) throws -> URL {
+        var components = URLComponents()
+        components.scheme = "https"
+        components.host = "api.codesquad.kr"
+        components.path = "/onban/\(type)"
+        
+        guard let url = components.url else {
+            throw NetworkError.wrongEndPoint
+        }
+        
+        return url
     }
     
     func fetchImageData(url: URL, completion: @escaping (Result<Data, NetworkError>) -> Void) {
